requirements.txt
# Core dependencies
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0
numpy>=1.24.0
scipy>=1.10.0
einops>=0.7.0

# Data processing
datasets>=2.15.0
sentencepiece>=0.1.99
tokenizers>=0.15.0
huggingface-hub>=0.19.0

# Training utilities
wandb>=0.16.0
hydra-core>=1.3.2
omegaconf>=2.3.0
accelerate>=0.25.0
tensorboard>=2.15.0

# Optimization
flash-attn>=2.3.3
triton>=2.1.0
bitsandbytes>=0.41.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.12.0
isort>=5.13.0
flake8>=6.1.0
mypy>=1.8.0

# CUDA extensions
pybind11>=2.11.0
ninja>=1.11.0

# Monitoring
psutil>=5.9.0
gputil>=1.4.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0
myst-parser>=2.0.0 